+++ test/data_table_2_data_column_2_test.dart
@@ -1308,7 +1308,7 @@ void main() {
             DataColumn2(
               label: const Center(child: Text('Name')),
               tooltip: 'Name',
-              onSort: sortEnabled ? (_, __) {} : null,
+              onSort: sortEnabled ? (_, _) {} : null,
             ),
           ],
           rows: const <DataRow2>[
@@ -1393,7 +1393,7 @@ void main() {
             DataColumn2(
               label: const Center(child: Text('column2')),
               tooltip: 'Column2',
-              onSort: (_, __) {},
+              onSort: (_, _) {},
             ),
           ],
           rows: const <DataRow2>[