+++ test/fixed_rows_cols_test.dart
@@ -14,7 +14,7 @@ import 'package:golden_toolkit/golden_toolkit.dart';
 import 'test_utils.dart';
 
 // Table with 10 rows, aproximately 450 pixel tall
-main() {
+void main() {
   group('Fixed colums/corner colors', () {
     testWidgets('0 cols, 0 rows', (WidgetTester tester) async {
       await wrapWidgetSetSurf(